+++ api/src/modules/authentication/dtos/signup-user-out.dto.ts
@@ -1,0 +1,10 @@
+import { Expose } from 'class-transformer';
+
+import { User } from '../../../modules/user/user.entity';
+
+export class SignupUserOutDto extends User {
+
+  @Expose()
+  requiresEmailValidation: boolean;
+
+}