+++ api/CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Return email requires validation
+
 ## [0.4.1] - 2020-07-11
 
 ### Changed